+++ model_navigator/framework_api/pipelines/builders/profiling.py
@@ -33,16 +33,9 @@ def profiling_builder(config: Config, package_descriptor: "PackageDescriptor") -
     for model_status in package_descriptor.navigator_status.model_status:
         for runtime_results in model_status.runtime_results:
             if runtime_results.status == Status.OK:
-                runner = package_descriptor.get_runner(
-                    format=model_status.format,
-                    jit_type=model_status.torch_jit,
-                    precision=model_status.precision,
-                    runtime=runtime_results.runtime,
-                )
                 commands.append(
                     Performance(
                         name=f"Performance {model_status.format.value}",
-                        runner=runner,
                         target_format=model_status.format,
                         requires=(load_metadata, load_samples),
                         target_jit_type=model_status.torch_jit,